Fixed issue in creating paymentService DB + fixed ticket price format (2 decimals)

+++ PaymentService/src/main/kotlin/it/polito/wa2/lab5/g09/paymentservice/utils/TransactionDeserializer.kt
@@ -30,7 +30,7 @@ data class TransactionInfo(
     @JsonProperty("orderId")
     val orderId: ObjectId,
     @JsonProperty("amount")
-    val amount: Float,
+    val amount: Double,
     @JsonProperty("creditCardNumber")
     val creditCardNumber: String,
     @JsonProperty("expirationDate")
